fix deadlock again
